requirements.txt
python-binance
python-telegram-bot
Flask
aiohttp
numpy
python-dotenv
supabase
telegram
python-dateutil
binance-connector
binance
